+++ esbuild.js
@@ -13,6 +13,7 @@ function copyAssets() {
     ["media/show-toast.ps1", "dist/show-toast.ps1"],
     ["media/focus-window.ps1", "dist/focus-window.ps1"],
     ["media/hide-toast.ps1", "dist/hide-toast.ps1"],
+    ["media/make-strips.ps1", "dist/make-strips.ps1"],
     ["media/icon.png", "dist/icon.png"],
   ];
   for (const [from, to] of assets) {